+++ backend/connection/dbConnection.js
@@ -2,8 +2,6 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 var mongoose = require("mongoose");
 var User = require("../models/User.model");
-var Message = require("../models/messages.model");
-var Invite = require("../models/Invites.model");
 var connection = "mongodb://localhost:27017/chatDB";
 var connectDb = function () {
     return mongoose.connect(connection);
