+++ backend/connection/dbConnection.ts
@@ -1,7 +1,5 @@
 import * as mongoose from "mongoose";
 const User = require("../models/User.model");
-const Message = require("../models/messages.model");
-const Invite = require("../models/Invites.model");
 
 const connection = "mongodb://localhost:27017/chatDB";
