+++ backend/connection/wsConnection.js
@@ -5,21 +5,10 @@ var io = require("socket.io")(4000, {
         origin: "*",
     },
 });
-var me = "";
-var you = "";
-var messages = [];
-if (messages.length >= 11)
-    messages.shift();
 io.on("connection", function (socket) {
-    socket.on("sender_reciever", function (_a) {
-        var sender = _a.sender, reciever = _a.reciever;
-        me = sender;
-        you = reciever;
-    });
-    socket.emit("send_message", { me: me, you: you });
-    // emiting and getting messages
     socket.on("message", function (_a) {
         var name = _a.name, message = _a.message;
+        console.log(message);
         io.emit("message", {
             name: name,
             message: message,
