--- backend/models/User.model copy.js
@@ -1,11 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var mongoose = require("mongoose");
-var UserSchema = new mongoose.Schema({
-    username: {
-        type: String,
-        required: true,
-    },
-});
-var User = mongoose.model("User", UserSchema);
-module.exports = User;
