+++ backend/request.rest
@@ -1,0 +1,1 @@
+GET http://localhost:4001/users/svetozar
