+++ backend/router/router.ts
@@ -4,37 +4,63 @@ const createError = require("http-errors");
 
 import { Request, Response } from "express";
 
-const userRoute = require("./users_route/users_route");
-const inviteRoute = require("./invite_route/invite_route");
-const Message = require("../models/messages.model");
-
-route.use("/", userRoute);
-route.use("/", inviteRoute);
+const User = require("../models/User.model");
 
 module.exports = route;
 
-// route.get("/hi/:sender/:reciever", async (req: Request, res: Response) => {
-//   try {
-//     const message = await Message.find({
-//       sender: req.params.sender,
-//       reciever: req.params.reciever,
-//     }).exec();
-//     res.json({ message: message });
-//   } catch (error) {
-//     res.json({ error: error });
-//   }
-// });
+// login auth
+route.get("/users/:username", async (req: Request, res: Response) => {
+  try {
+    const users = await User.findOne({ username: req.params.username }).exec();
+    if (!users || undefined)
+      throw createError(
+        400,
+        "Invalid input",
+        `User ${req.params.username} doesnt exist`,
+      );
+    res.json({ message: users });
+  } catch (error: any) {
+    res.status(error.status);
+    res.json({ errorStatus: error.status, message: error, stack: error.stack });
+  }
+});
 
-// route.post("/:sender/:reciever", async (req: Request, res: Response) => {
-//   try {
-//     const message = new Message({
-//       sender: req.params.sender,
-//       reciever: req.params.reciever,
-//       messages: req.body.message,
-//     });
-//     await message.save();
-//     res.json({ message: message });
-//   } catch (error) {
-//     res.json({ error: "error" });
-//   }
-// });
+// create new users
+
+route.post("/register", async (req: Request, res: Response) => {
+  try {
+    const users = await User.find().exec();
+    for (let i = 0; i < users.length; i++) {
+      if (users[i].username === req.body.username) {
+        throw createError(
+          400,
+          "Invalid input",
+          `User ${req.body.username} already exist`,
+        );
+      }
+    }
+    const user = new User({ type: "POST", username: req.body.username });
+    if (req.body.username === "")
+      throw createError(400, "Invalid input", `Please enter valid input`);
+    await user.save();
+    res.status(201).send({ message: `User ${req.body.username} created` }); //ok response and creating
+  } catch (error: any) {
+    res.status(error.status);
+    res.json({ errorStatus: error.status, message: error });
+  }
+});
+
+// delete users
+
+route.delete("/:username", (req: Request, res: Response) => {
+  try {
+    User.deleteOne({ username: req.params.username }).exec();
+    res
+      .json({
+        message: `deleted user: ${req.params.username}`,
+      })
+      .sendStatus(204);
+  } catch (error) {
+    res.json({ error }).sendStatus(501);
+  }
+});
