--- backend/router/users_route/users_route.js
@@ -1,146 +1,0 @@
-"use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (_) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-var express = require("express");
-var route = express.Router();
-var createError = require("http-errors");
-var User = require("../../models/User.model");
-route.get("/users", function (req, res) { return __awaiter(void 0, void 0, void 0, function () {
-    var users, error_1;
-    return __generator(this, function (_a) {
-        switch (_a.label) {
-            case 0:
-                _a.trys.push([0, 2, , 3]);
-                return [4 /*yield*/, User.find({}).exec()];
-            case 1:
-                users = _a.sent();
-                if (!users || undefined)
-                    throw createError(404, "No users found");
-                res.json({ users: users });
-                return [3 /*break*/, 3];
-            case 2:
-                error_1 = _a.sent();
-                res.status(error_1.status);
-                res.json({
-                    errorStatus: error_1.status,
-                    message: error_1,
-                });
-                return [3 /*break*/, 3];
-            case 3: return [2 /*return*/];
-        }
-    });
-}); });
-route.get("/users/:username", function (req, res) { return __awaiter(void 0, void 0, void 0, function () {
-    var users, error_2;
-    return __generator(this, function (_a) {
-        switch (_a.label) {
-            case 0:
-                _a.trys.push([0, 2, , 3]);
-                return [4 /*yield*/, User.findOne({
-                        username: req.params.username,
-                    }).exec()];
-            case 1:
-                users = _a.sent();
-                if (!users || undefined)
-                    throw createError(400, "Invalid input", "User " + req.params.username + " doesnt exist");
-                res.json({ message: users });
-                return [3 /*break*/, 3];
-            case 2:
-                error_2 = _a.sent();
-                res.status(error_2.status);
-                res.json({ errorStatus: error_2.status, message: error_2, stack: error_2.stack });
-                return [3 /*break*/, 3];
-            case 3: return [2 /*return*/];
-        }
-    });
-}); });
-route.post("/register", function (req, res) { return __awaiter(void 0, void 0, void 0, function () {
-    var users, i, user, error_3;
-    return __generator(this, function (_a) {
-        switch (_a.label) {
-            case 0:
-                _a.trys.push([0, 3, , 4]);
-                return [4 /*yield*/, User.find().exec()];
-            case 1:
-                users = _a.sent();
-                for (i = 0; i < users.length; i++) {
-                    if (users[i].username === req.body.username) {
-                        throw createError(400, "Invalid input", "User " + req.body.username + " already exist");
-                    }
-                }
-                user = new User({ type: "POST", username: req.body.username });
-                if (req.body.username === "")
-                    throw createError(400, "Invalid input", "Please enter valid input");
-                return [4 /*yield*/, user.save()];
-            case 2:
-                _a.sent();
-                res.status(201).send({ message: "User " + req.body.username + " created" }); //ok response and creating
-                return [3 /*break*/, 4];
-            case 3:
-                error_3 = _a.sent();
-                res.status(error_3.status);
-                res.json({ errorStatus: error_3.status, message: error_3 });
-                return [3 /*break*/, 4];
-            case 4: return [2 /*return*/];
-        }
-    });
-}); });
-route.delete("/:username", function (req, res) { return __awaiter(void 0, void 0, void 0, function () {
-    var data_1, error_4;
-    return __generator(this, function (_a) {
-        switch (_a.label) {
-            case 0:
-                _a.trys.push([0, 2, , 3]);
-                return [4 /*yield*/, User.deleteOne({ username: req.params.username }).exec()];
-            case 1:
-                data_1 = _a.sent();
-                res
-                    .json({
-                    message: "deleted user: " + req.params.username,
-                })
-                    .sendStatus(204);
-                return [3 /*break*/, 3];
-            case 2:
-                error_4 = _a.sent();
-                res.json({ error: error_4 }).sendStatus(501);
-                return [3 /*break*/, 3];
-            case 3: return [2 /*return*/];
-        }
-    });
-}); });
-module.exports = route;
