+++ backend/server.js
@@ -10,9 +10,17 @@ var data = require("./router/router");
 app.use(cors());
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
+app.use(bodyParser.urlencoded({
+    extended: true,
+}));
 connectDb();
 // routes importing
 app.use("/", data);
+app.use(function (err, req, res) {
+    res.locals.error = err;
+    var status = err.status || 500;
+    res.status(status);
+});
 var port = 4001;
 app.listen(port, function () {
     console.log("listening on port " + port);
