+++ frontend/package.json
@@ -19,7 +19,6 @@
     "next-cookies": "^2.0.3",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-icons": "^4.2.0",
     "socket.io-client": "^4.2.0"
   },
   "devDependencies": {
