+++ frontend/pages/_app.tsx
@@ -1,18 +1,16 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
 import { useCookie } from "next-cookie";
-import { GetServerSideProps, NextPage } from "next";
-import { FC } from "react";
+import { GetServerSideProps } from "next";
 
-const MyApp = ({ Component, pageProps }: AppProps, props: string) => {
-  const cookie = useCookie(props.cookie);
+function MyApp({ Component, pageProps }: AppProps) {
   return (
     <>
       <div className="BIG"></div>
-      <Component {...pageProps} Cookie={cookie} />
+      <Component {...pageProps} />
     </>
   );
-};
+}
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
   const cookie = useCookie(context);
